Skip failing tests on Perl >= 5.33.5
The new stat() implementation introduced in Perl 5.33.5 sets dev, ino and
rdev differently. There is no point in altering alt_stat() to match it for
new versions of Perl given that this whole module is redundant now since
the new stat() and utime() (and lstat()) handle UTC file times correctly
anyway, so just skip the tests on new versions of Perl instead.